+++ src/google/codetablewriter_interface.h
@@ -56,7 +56,7 @@ class CodeTableWriterInterface {
   virtual void Add(const char* data, size_t size) = 0;
 
   // Encode a COPY opcode with args "offset" (into dictionary) and "size" bytes.
-  virtual void Copy(int32_t offset, size_t size) = 0;
+  virtual void Copy(int64_t offset, size_t size) = 0;
 
   // Encode a RUN opcode for "size" copies of the value "byte".
   virtual void Run(size_t size, unsigned char byte) = 0;
